Extending tests for clear steps and failed results
