+++ pi4j-test/src/main/java/com/pi4j/test/About.java
@@ -29,7 +29,7 @@ import com.pi4j.exception.Pi4JException;
 import com.pi4j.io.IOType;
 import com.pi4j.platform.Platform;
 import com.pi4j.provider.Provider;
-import com.pi4j.test.smokeTest.ProviderContext;
+import com.pi4j.test.smoketest.ProviderContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
