+++ pi4j-test/src/main/java/com/pi4j/test/Main.java
@@ -24,7 +24,7 @@ package com.pi4j.test;/*-
  */
 
 import com.pi4j.io.IOType;
-import com.pi4j.test.smokeTest.*;
+import com.pi4j.test.smoketest.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
