+++ pi4j-test/src/main/java/com/pi4j/test/smoketest/TestResult.java
@@ -1,4 +1,4 @@
-package com.pi4j.test.smokeTest;
+package com.pi4j.test.smoketest;
 
 public record TestResult(String name, boolean success, String message) {
     public String getLogOutput() {
